requirements.txt
streamlit
opencv-python
mediapipe==0.10.14
numpy
diskcache
piper-tts
qdrant-client
pyttsx3
protobuf>=4.25.3,<5
python-dotenv
tqdm
